sciagent-rag/backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "AI interactive science simulation backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "jsonschema>=4.26.0",
    "langchain>=1.2.12",
    "langchain-anthropic>=0.3.14",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=0.2.27",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.7",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]

[tool.setuptools]
py-modules = [
    "main",
    "rag_agent",
    "multimodal_ocr_manager",
    "simulation_command_validator",
    "config",
    "command_prompt_manager",
    "langgraph_manager",
    "llm_factory",
    "vector_db_manager",
]
